index default export 하도록 수정
